Split auth API user to its own file
